+++ PlaylistCompletionStatusOption.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Playlist
+{
+    public sealed class PlaylistCompletionStatusOption
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; }
+    }
+}